Added JUnit test for dice.roll()
