+++ src/terning/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author A
+ *
+ */
+package terning;
